Update 1.把java 文件改成kotlin 文件。 2.修改kotlin 静态方法。 记录一下，kotlin 切分字符串无需转义符。

+++ library_common/src/main/java/com/phone/library_common/manager/TextViewStyleManager.kt
@@ -66,7 +66,7 @@ object TextViewStyleManager {
         spannable.setSpan(
             AbsoluteSizeSpan(spToPx(size)),
             start,
-            end.toInt(),
+            end,
             Spannable.SPAN_INCLUSIVE_INCLUSIVE
         )
         //SpannableString字体大小不一致垂直居中显示
